backend/pyproject.toml
[tool.poetry]
name = "gen_ui_backend"
version = "0.1.0"
description = ""
authors = ["Connor Haines <connor_haines@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-core = "^0.3.10"
langgraph = "^0.2.35"
langserve = "^0.3.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.31.1"
fastapi = "^0.115.0"
pydantic = ">=2.8.2,<2.9.0" 
langchain-openai = "^0.2.2"
langchain-community = "^0.3.2"
typer = "^0.12.5"
sse-starlette = "^2.1.3"
clerk-backend-api = "^1.0.3"
psycopg2-binary = "^2.9.10"
sqlmodel = "^0.0.22"
sqlalchemy = "^2.0.36"
psycopg = "^3.2.3"
psycopg-pool = "^3.2.3"
langgraph-checkpoint-postgres = "^2.0.2"
psycopg-binary = "^3.2.3"


[tool.poetry.scripts]
start = "gen_ui_backend.server:start"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
google-cloud-storage = "^2.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
